+++ spp_alerts/tests/__init__.py
@@ -3,3 +3,4 @@ from . import common
 from . import test_alert
 from . import test_alert_rule
 from . import test_rule_evaluation
+from . import test_rule_evaluation_access